refactor(core): extract proximadb-search-types foundation crate
Move the self-contained, foundation-tier leaves of src/core/search into a
new crates/foundation/proximadb-search-types crate and re-export them from
the root as `pub use proximadb_search_types::{...}`, so every existing
`crate::core::search::{results, bounded_queue, sql_value_filter,
json_value_serde, json_comparison}::*` path resolves unchanged.

Behaviour-neutral extraction, part of the root-crate decomposition
(Slice D / D1) shrinking the monolithic root compilation unit to fix CI
OOMs.

Extracted:
- results.rs, bounded_queue.rs, sql_value_filter.rs, json_value_serde.rs
  (whole-file moves)
- json_comparison (carved out of the inline `pub mod json_comparison` block
  in search/mod.rs into its own file)

Mechanical details:
- rewrite upward refs to foundation crates: distance_computation ->
  proximadb_distance_kernel, core::search::{ComparisonOperator,
  FilterExpression} -> proximadb_filter_expression, crate::proto ->
  proximadb_proto; intra-crate refs -> crate::{results,sql_value_filter,
  json_comparison}
- move `impl From<OptimizedSearchRecord> for SearchVectorRecord` (and the
  &-variant) out of core::conversions into results.rs: now that
  OptimizedSearchRecord is foreign to root, the orphan rule forbids the impl
  there; it is legal in the crate where the type is local. The root proto
  envelope fn keeps working (trait impls are global).
- new Cargo.toml ([lib] name = proximadb_search_types) + workspace wiring

Not extracted (stay in root): UnifiedSearchParams (3 upward edges: compute
DistanceMetric, modality UnifiedQuantizationLevel, query
FilterOptimizationHints) and the inline simple enums interleaved with it.

Verified: cargo check -p proximadb --tests, clippy --lib --bins -D,
server/bench/migrate bins, nextest -p proximadb-search-types (17/17),
layering green, capability-matrix validator green.
