+++ crates/foundation/proximadb-search-types/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "proximadb-search-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Foundation search-result and filter-value types for ProximaDB: search-result records, bounded result queue, SQL/JSON value filtering, and JSON value (de)serialization"
+
+[lib]
+name = "proximadb_search_types"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+proximadb-proto.workspace = true
+proximadb-kernel.workspace = true
+proximadb-records.workspace = true
+proximadb-data-model.workspace = true
+proximadb-filter-expression.workspace = true
+proximadb-distance-kernel.workspace = true
+
+[lints]
+workspace = true
