+++ docs/12-design/ROOT_CRATE_DECOMPOSITION_SLICE_D_PORT_INVERSION_2026_06_26.adoc
@@ -145,8 +145,8 @@ then depends *down* on the trait crate, never *up* on the implementor.
 [cols="1,4,1",options="header"]
 |===
 | Phase | Work | Risk
-| D0 | **Redirect-down (Mechanism A, existing crates).** Rewrite storage's `crate::compute::…DistanceMetric` / `crate::core::search::FilterExpression` / quant-types / config / hardware-caps / index-types references to the existing foundation crates. Pure path rewrites; verify each root path == leaf type. Several small mechanical PRs; collapses the edge count first. | Low
-| D1 | **Extract remaining shared-type leaves down.** `proximadb-search-types` (← `core::search` minus FilterExpression) and `proximadb-bloom` (← `core::bloom`). Same leaf recipe already proven. | Low
+| D0 | **Redirect-down (Mechanism A, existing crates).** Rewrite storage's `crate::compute::…DistanceMetric` / `crate::core::search::FilterExpression` / quant-types / config / hardware-caps / index-types references to the existing foundation crates. Pure path rewrites; verify each root path == leaf type. Several small mechanical PRs; collapses the edge count first. **DONE:** distance-kernel + FilterExpression (PR #370), hardware-caps + index-types (D0b, PR #377). Remaining: quant-types (unverified), config (divergent root module — not a clean redirect). | Low
+| D1 | **Extract remaining shared-type leaves down.** `proximadb-search-types` (← `core::search` minus FilterExpression) and `proximadb-bloom` (← `core::bloom`). Same leaf recipe already proven. **DONE:** `proximadb-bloom` (PR #379); `proximadb-search-types` — the whole-file leaves `results`/`bounded_queue`/`sql_value_filter`/`json_value_serde` + the carved-out `json_comparison` module (this PR). `UnifiedSearchParams` stays in root (3 upward edges: compute `DistanceMetric`, modality `UnifiedQuantizationLevel`, query `FilterOptimizationHints`); the inline simple enums (`SearchMode`/`BlockPruneConfig`/…) remain interleaved with it and are deferred. | Low
 | D2 | **Define `proximadb-storage-ports`** — the `IndexPort` + `CollectionMetadataPort` traits (and the conditional distance/quant dispatch ports if the decision rule says invert). Trait-only crate, no impls. | Low
 | D3 | **Refactor storage to consume injected ports.** `engine.rs`/compaction stop constructing `AxisManager` / calling `CollectionService` directly; take `Arc<dyn IndexPort>` / `Arc<dyn CollectionMetadataPort>` via constructor injection. Same concrete behavior, now behind the trait. **This is the hot, coordination-window phase.** | Med-High
 | D4 | **Implement ports + wire injection.** `impl IndexPort for AxisManager`, `impl CollectionMetadataPort for CollectionService`; the composition root injects them. | Med
